+++ Main.cpp
@@ -1,7 +1,6 @@
 #include "Numbers.h"
 #include <iostream>
 #include <fstream>
-
 using namespace std;
 namespace simple_shapes {
 	void Init(container &c);
@@ -11,12 +10,15 @@ namespace simple_shapes {
 	polcoor* InPolcoor(polcoor &l, ifstream &ifst);
 	Node* In(ifstream &ifst);
 	void In(container &c, ifstream &ifst);
-	void OutComplex(complex *p, ofstream &ofst);
 	void Out(container &c, ofstream &ofst);
 	void Out(Node *s, ofstream &ofst);
 	int addlist(container &c, ifstream &ifst);
 	void OutSimple(simple *t, ofstream &ofst);
+	void OutComplex(complex *p, ofstream &ofst);
 	void OutPolcoor(polcoor *l, ofstream &ofst);
+	float numbers_s(numbers *s);
+	bool Compare(numbers *p, numbers *t);
+	void Sort(container &c);
 	void Out_only_complex(container &c, ofstream &ofst);
 }
 using namespace simple_shapes;
@@ -38,9 +40,15 @@ int main(int argc, char* argv[])
 	In(c, ifst);
 	ofst << "Filled container. " << endl;
 
+	Out(c, ofst);
+	Sort(c);
+	ofst << endl;
+	Out(c, ofst);
+	ofst << endl;
 	Out_only_complex(c, ofst);
-
+	ofst << endl;
 	Clear(c);
+
 	ofst << "Empty container. " << endl;
 	Out(c, ofst);
 	cout << "Stop" << endl;
