+++ filein.txt
@@ -1,12 +1,24 @@
-3
-15 36 סל
 2
 12 18 סל
 2
-24 5 הל
+24 5 לל
 1
-12 169 לל
+12 169 ל 
 1
-12 -169 ל
-
-
+12 -169 הל
+3
+12 49
+3
+4 44
+2
+12 18 סל
+2
+24 5 לל
+1
+12 169 ל 
+1
+12 -169 הל
+3
+12 49
+3
+4 44
