+++ fileout.txt
@@ -1,0 +1,101 @@
+Filled container. 
+ Container contains 12 elements.
+0: It is Simple number: числитель = 12, знаменатель = 18
+12/18, еденица измерений - см
+значение = 0.666667
+1: It is Simple number: числитель = 24, знаменатель = 5
+24/5, еденица измерений - мм
+значение = 4.8
+2: It is Complex number: действительная часть = 12, мнимая часть = 169
+12+169i, еденица измерений - м
+значение = 169.426
+3: It is Complex number: действительная часть = 12, мнимая часть = -169
+12-169i, еденица измерений - дм
+значение = 169.426
+4: It is Polar coordinates number: угол = 12, расстояние до точки = 49
+(49,12°)
+значение = 15.0568
+5: It is Polar coordinates number: угол = 4, расстояние до точки = 44
+(44,4°)
+значение = -62.0596
+6: It is Simple number: числитель = 12, знаменатель = 18
+12/18, еденица измерений - см
+значение = 0.666667
+7: It is Simple number: числитель = 24, знаменатель = 5
+24/5, еденица измерений - мм
+значение = 4.8
+8: It is Complex number: действительная часть = 12, мнимая часть = 169
+12+169i, еденица измерений - м
+значение = 169.426
+9: It is Complex number: действительная часть = 12, мнимая часть = -169
+12-169i, еденица измерений - дм
+значение = 169.426
+10: It is Polar coordinates number: угол = 12, расстояние до точки = 49
+(49,12°)
+значение = 15.0568
+11: It is Polar coordinates number: угол = 4, расстояние до точки = 44
+(44,4°)
+значение = -62.0596
+
+ Container contains 12 elements.
+0: It is Complex number: действительная часть = 12, мнимая часть = 169
+12+169i, еденица измерений - м
+значение = 169.426
+1: It is Complex number: действительная часть = 12, мнимая часть = -169
+12-169i, еденица измерений - дм
+значение = 169.426
+2: It is Complex number: действительная часть = 12, мнимая часть = 169
+12+169i, еденица измерений - м
+значение = 169.426
+3: It is Complex number: действительная часть = 12, мнимая часть = -169
+12-169i, еденица измерений - дм
+значение = 169.426
+4: It is Polar coordinates number: угол = 12, расстояние до точки = 49
+(49,12°)
+значение = 15.0568
+5: It is Polar coordinates number: угол = 12, расстояние до точки = 49
+(49,12°)
+значение = 15.0568
+6: It is Simple number: числитель = 24, знаменатель = 5
+24/5, еденица измерений - мм
+значение = 4.8
+7: It is Simple number: числитель = 24, знаменатель = 5
+24/5, еденица измерений - мм
+значение = 4.8
+8: It is Simple number: числитель = 12, знаменатель = 18
+12/18, еденица измерений - см
+значение = 0.666667
+9: It is Simple number: числитель = 12, знаменатель = 18
+12/18, еденица измерений - см
+значение = 0.666667
+10: It is Polar coordinates number: угол = 4, расстояние до точки = 44
+(44,4°)
+значение = -62.0596
+11: It is Polar coordinates number: угол = 4, расстояние до точки = 44
+(44,4°)
+значение = -62.0596
+
+Only complex.
+0: It is Complex number: действительная часть = 12, мнимая часть = 169
+12+169i, еденица измерений - м
+значение = 169.426
+1: It is Complex number: действительная часть = 12, мнимая часть = -169
+12-169i, еденица измерений - дм
+значение = 169.426
+2: It is Complex number: действительная часть = 12, мнимая часть = 169
+12+169i, еденица измерений - м
+значение = 169.426
+3: It is Complex number: действительная часть = 12, мнимая часть = -169
+12-169i, еденица измерений - дм
+значение = 169.426
+4: 
+5: 
+6: 
+7: 
+8: 
+9: 
+10: 
+11: 
+
+Empty container. 
+ Container contains 0 elements.
